optimize with queries

+++ db/migrate/202507180832_performance_indexs.rb
@@ -1,12 +1,19 @@
 class PerformanceIndexs < ActiveRecord::Migration[5.1]
   def up
-    add_index :journals, [:journalized_id, :journalized_type, :user_id, :private_notes],
-              name: 'index_journals_for_relevant_lookup'
-    add_index :level_agreement_policies, :project_id
+    unless index_exists?(:journals, [:journalized_id, :journalized_type, :user_id, :private_notes], name: 'index_journals_for_relevant_lookup')
+      add_index :journals, [:journalized_id, :journalized_type, :user_id, :private_notes], name: 'index_journals_for_relevant_lookup'
+    end
+    unless index_exists?(:level_agreement_policies, :project_id)
+      add_index :level_agreement_policies, :project_id
+    end
   end
 
   def down
-    remove_index :journals, name: 'index_journals_for_relevant_lookup'
-    remove_index :level_agreement_policies, :project_id
+    if index_exists?(:journals, [:journalized_id, :journalized_type, :user_id, :private_notes], name: 'index_journals_for_relevant_lookup')
+      remove_index :journals, name: 'index_journals_for_relevant_lookup'
+    end
+    if index_exists?(:level_agreement_policies, :project_id)
+      remove_index :level_agreement_policies, :project_id
+    end
   end
 end
